+++ src/modules/works/work.controller.ts
@@ -1,11 +1,24 @@
-import { Controller, Get, Post, Body } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Put,
+  Query,
+  Delete,
+  Param,
+} from '@nestjs/common';
 import { WorkService } from './work.service';
 import { WorksDto } from './work.dto';
 
-@Controller('/works')
+@Controller('works')
 export class WorkController {
   constructor(private readonly workService: WorkService) {}
 
+  @Get(':id')
+  getWork(@Param('id') id: string) {
+    return this.workService.getWork(id);
+  }
   @Get()
   getWorks() {
     return this.workService.getWorks();
@@ -14,5 +27,15 @@ export class WorkController {
   @Post()
   createWork(@Body() dto: WorksDto) {
     return this.workService.createWork(dto);
+  }
+
+  @Put()
+  updateWork(@Query('id') id: string, @Body() dto: WorksDto) {
+    return this.workService.updateWork(id, dto);
+  }
+
+  @Delete()
+  deleteWork(@Query('id') id: string) {
+    return this.workService.deleteWork(id);
   }
 }